+++ src/test/java/tools/devnull/kodo/ExpectationTests.java
@@ -42,7 +42,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static tools.devnull.kodo.Expectation.performing;
 import static tools.devnull.kodo.Expectation.it;
 import static tools.devnull.kodo.Expectation.the;
 import static tools.devnull.kodo.Expectation.to;
@@ -184,7 +183,7 @@ public class ExpectationTests {
     Function function = mock(Function.class);
     when(function.apply(value)).thenReturn(value);
 
-    performing(function).accept(value);
+    the(function).accept(value);
 
     verify(function).apply(value);
   }