+++ src/test/java/tools/devnull/kodo/SpecTests.java
@@ -73,6 +73,8 @@ public class SpecTests {
   private Predicate failTest;
   @Mock
   private Function function;
+  @Mock
+  private Function<Object, Boolean> booleanFunction;
 
   @Before
   public void initialize() {
@@ -80,6 +82,7 @@ public class SpecTests {
     when(failTest.test(anyObject())).thenReturn(false);
     when(function.apply(target)).thenReturn(value);
     doThrow(exception).when(failOperation).accept(anyObject());
+    when(booleanFunction.apply(target)).thenReturn(true);
   }
 
   @Test
@@ -89,6 +92,12 @@ public class SpecTests {
 
     assertSame(spec, spec.expect(operation, test));
     assertSame(spec, spec.expect(operation, test, message));
+
+    assertSame(spec, spec.expect(booleanFunction));
+    assertSame(spec, spec.expect(booleanFunction, message));
+
+    assertSame(spec, spec.expect(true));
+    assertSame(spec, spec.expect(true, message));
 
     assertSame(spec, spec.when(operation));
     assertSame(spec, spec.when(runnableOperation));
@@ -114,19 +123,46 @@ public class SpecTests {
   }
 
   @Test
-  public void testThenWithFunction() {
+  public void testExpectWithFunction() {
     spec.expect(function, test);
     verify(function).apply(target);
     verify(test).test(value);
   }
 
   @Test
-  public void testThenWithConsumer() {
+  public void testExpectWithConsumer() {
     spec.expect(operation, test);
     verify(operation).accept(target);
 
     spec.expect(failOperation, test);
     verify(test).test(exception);
+  }
+
+  @Test
+  public void testExpectWithBoolean() {
+    spec.expect(booleanFunction);
+
+    verify(booleanFunction).apply(target);
+  }
+
+  @Test(expected = AssertionError.class)
+  public void testFailExpectWithBoolean() {
+    spec.expect(false);
+  }
+
+  @Test(expected = AssertionError.class)
+  public void testFailExpectWithBooleanFunction() {
+    spec.expect(o -> false);
+  }
+
+  @Test
+  public void testSucceedExpectWithBoolean() {
+    spec.expect(true);
+  }
+
+  @Test
+  public void testSucceedExpectWithBooleanFunction() {
+    spec.expect(o -> true);
   }
 
   @Test
@@ -147,6 +183,10 @@ public class SpecTests {
   public void testMessages() {
     assertMessage(() -> spec.expect(function, failTest, message));
     assertMessage(() -> spec.expect(operation, failTest, message));
+
+    when(booleanFunction.apply(target)).thenReturn(false);
+    assertMessage(() -> spec.expect(booleanFunction, message));
+    assertMessage(() -> spec.expect(false, message));
   }
 
   @Test